requirements.txt
absl-py==1.4.0
astor==0.8.1
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
cycler==0.11.0
flatbuffers==24.3.25
fonttools==4.38.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
hickle==5.0.2
idna==3.10
importlib-metadata==6.6.0
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==18.1.1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.5.3
mock==5.0.2
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pip==18.1
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.19.2
scipy==1.7.3
setuptools==40.6.2
six==1.16.0
tensorboard==1.13.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing-extensions==4.5.0
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.40.0
wrapt==1.16.0
zipp==3.15.0
